src/orchestrator/booking_agent/pyproject.toml
[project]
name = "booking-agent"
version = "1.0.0"
description = "AI-powered Order Booking Agent Microservice"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "litellm>=1.50.0",
    "python-dotenv>=1.0.0",
    "ioa-observe-sdk>=0.3.5",
    "a2a-sdk>=0.3.0",
    "agntcy-app-sdk>=0.4.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "agent", "domain", "config", "services"]
